refactor: move CorePage components into ThemePage
