+++ src/components/features/ThemePage/components/Badges/Badges.tsx
@@ -2,8 +2,8 @@ import { CoreSectionWrapper } from "../CoreSectionWrapper/CoreSectionWrapper";
 import { Card } from "@/components/core/Card/Card";
 import { Badge } from "@/components/core/Badge/Badge";
 import { Text } from "@/components/core/Text/Text";
-import { Flex, HStack, VStack } from "../../../../../styled-system/jsx";
-import { css } from "../../../../../styled-system/css";
+import { Flex, HStack, VStack } from "../../../../../../styled-system/jsx";
+import { css } from "../../../../../../styled-system/css";
 
 export const Badges = () => {
   return (
