+++ src/components/features/ThemePage/components/Inputs/Inputs.tsx
@@ -1,12 +1,10 @@
-"use client";
-
-import { useState, ChangeEvent } from "react";
+import { useState } from "react";
 import { CoreSectionWrapper } from "../CoreSectionWrapper/CoreSectionWrapper";
 import { Card } from "@/components/core/Card/Card";
 import { Checkbox } from "@/components/core/Checkbox/Checkbox";
 import { TextInput } from "@/components/core/TextInput/TextInput";
 import { TextArea } from "@/components/core/TextArea/TextArea";
-import { HStack, VStack } from "../../../../../styled-system/jsx";
+import { HStack, VStack } from "../../../../../../styled-system/jsx";
 import { Text } from "@/components/core/Text/Text";
 
 export function Inputs() {
